chore: reorganize some classes
