+++ lib/common/alarm/action/MultipleAlarmActionStrategy.ts
