+++ lib/common/alarm/action/NoopAlarmActionStrategy.ts
