+++ lib/common/alarm/index.ts
@@ -1,11 +1,7 @@
+export * from "./action";
 export * from "./AlarmFactory";
 export * from "./AlarmNamingStrategy";
 export * from "./CustomAlarmThreshold";
-export * from "./IAlarmActionStrategy";
 export * from "./IAlarmAnnotationStrategy";
 export * from "./IAlarmDedupeStringProcessor";
 export * from "./IAlarmNamingStrategy";
-export * from "./MultipleAlarmActionStrategy";
-export * from "./NoopAlarmActionStrategy";
-export * from "./OpsItemAlarmActionStrategy";
-export * from "./SnsAlarmActionStrategy";
