+++ lib/common/metric/MetricWithAlarmSupport.ts
@@ -3,6 +3,6 @@ import { MathExpression, Metric } from "aws-cdk-lib/aws-cloudwatch";
 /**
  * Any metric we can create an alarm on.
  *
- * (Cannot be IMetric, as it does not have any alarm support in general.)
+ * Cannot be an IMetric, as it does not have support for alarms.
  */
 export type MetricWithAlarmSupport = Metric | MathExpression;
